# Apply optimizations to speed up computation by 10x