--- a/gpt2.py
+++ b/gpt2.py
@@ -1,3 +1,4 @@
+import math
 from dataclasses import dataclass
 
 import torch
@@ -29,6 +30,7 @@ class CausalSelfAttention(nn.Module):
         # config
         self.n_head = config.n_head
         self.n_embd = config.n_embd
+        self.register_buffer("bias", torch.tril(torch.ones(config.block_size, config.block_size)).view(1, 1, config.block_size, config.block_size))
     
     def forward(self, x):
         B, T, C = x.size()  # batch_size, seq_len, n_embd
@@ -39,7 +41,16 @@ class CausalSelfAttention(nn.Module):
         q = q.view(B, T, self.n_head, C//self.n_head).transpose(1,2)    # (B, nh, T, hs) # transpose n_head with seq_len for attention calculation
         k = k.view(B, T, self.n_head, C//self.n_head).transpose(1,2)    # (B, nh, T, hs)
         v = v.view(B, T, self.n_head, C//self.n_head).transpose(1,2)    # (B, nh, T, hs)
+        
+        # raw attention
+        # att = (q @ k.transpose(-2, -1)) / math.sqrt(k.size(-1))  # (B, nh, T, T)
+        # att = att.masked_fill(self.bias[:,:,:T,:T] == 0, float('-inf'))
+        # att = F.softmax(att, dim=-1)
+        # y = att @ v  # (B, nh, T, hs)
+
+        # flash attention
         y = F.scaled_dot_product_attention(q, k, v, is_causal=True)     # flash attention # (B, nh, T, hs)
+
         y = y.transpose(1, 2).contiguous().view(B, T, C) # re-assemble all head outputs side by side   # (B, T, C)
         # output projection
         y = self.c_proj(y)  # (B, T, C)