Speed test:
- GPU vanila    | dt: 379.91ms | tok/sec: 43126.52
- TF32          | dt: 170.49ms | tok/sec: 96098.66
- BF16 mixed precision | dt: 161.31ms | tok/sec: 101567.53
- Torch compile | dt: 58.22ms | tok/sec: 281415.76
- Flash attention | dt: 43.76ms | tok/sec: 374403.29
- nice vocab size | dt: 40.78ms | tok/sec: 401729.68

**~10x speed up!!!!**
380ms --> 40 ms      
43k tps --> 401k tps

GPU config:
<img width="826" alt="image" src="https://github.com/user-attachments/assets/989ba5ab-9726-4d99-8d94-c9c39653a340" />